feat(misc): export mock builder context from workspace
