+++ packages/web/src/utils/testing.ts
@@ -1,17 +1,12 @@
 import { join } from 'path';
-import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
+
+import { schema } from '@angular-devkit/core';
 import { Rule, Tree } from '@angular-devkit/schematics';
-import {
-  BuilderContext,
-  Architect,
-  Target,
-  ScheduleOptions
-} from '@angular-devkit/architect';
-import {
-  TestingArchitectHost,
-  TestLogger
-} from '@angular-devkit/architect/testing';
-import { schema, JsonObject } from '@angular-devkit/core';
+import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
+import { Architect } from '@angular-devkit/architect';
+import { TestingArchitectHost } from '@angular-devkit/architect/testing';
+
+import { MockBuilderContext } from '@nrwl/workspace/testing';
 
 const testRunner = new SchematicTestRunner(
   '@nrwl/web',
@@ -44,75 +39,4 @@ export async function getMockContext() {
   const context = new MockBuilderContext(architect, architectHost);
   await context.addBuilderFromPackage(join(__dirname, '../..'));
   return context;
-}
-
-export class MockBuilderContext implements BuilderContext {
-  id: 0;
-
-  builder: any = {};
-  analytics = null;
-
-  target: Target = {
-    project: null,
-    target: null
-  };
-
-  get currentDirectory() {
-    return this.architectHost.currentDirectory;
-  }
-  get workspaceRoot() {
-    return this.architectHost.workspaceRoot;
-  }
-  logger = new TestLogger('test');
-  constructor(
-    private architect: Architect,
-    private architectHost: TestingArchitectHost
-  ) {}
-
-  async addBuilderFromPackage(path: string) {
-    return await this.architectHost.addBuilderFromPackage(path);
-  }
-
-  async addTarget(target: Target, builderName: string) {
-    return await this.architectHost.addTarget(target, builderName);
-  }
-
-  getBuilderNameForTarget(target: Target) {
-    return this.architectHost.getBuilderNameForTarget(target);
-  }
-
-  scheduleTarget(
-    target: Target,
-    overrides?: JsonObject,
-    scheduleOptions?: ScheduleOptions
-  ) {
-    return this.architect.scheduleTarget(target, overrides, scheduleOptions);
-  }
-
-  scheduleBuilder(
-    name: string,
-    overrides?: JsonObject,
-    scheduleOptions?: ScheduleOptions
-  ) {
-    return this.architect.scheduleBuilder(name, overrides, scheduleOptions);
-  }
-
-  getTargetOptions(target: Target) {
-    return this.architectHost.getOptionsForTarget(target);
-  }
-
-  validateOptions<T extends JsonObject = JsonObject>(
-    options: JsonObject,
-    builderName: string
-  ): Promise<T> {
-    return Promise.resolve<T>(options as T);
-  }
-
-  reportRunning() {}
-
-  reportStatus(status: string) {}
-
-  reportProgress(current: number, total?: number, status?: string) {}
-
-  addTeardown(teardown: () => Promise<void> | void) {}
 }
